reward_models/llava-server/requirements.txt
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bert-score==0.3.13
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.0
contourpy==1.1.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
fibonacci==1.0.0
filelock==3.12.2
flash-attn==2.5.8
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.23.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
linkify-it-py==2.0.2
lit==16.0.6
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.7.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1.3
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
orjson==3.9.10
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.1
psutil==5.9.5
py-cpuinfo==9.0.0
py-fibonacci==0.5.2
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.32.3
rpds-py==0.13.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.37.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.28.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.0.147
urllib3==2.1.0
uvicorn==0.24.0.post1
wandb==0.16.0
wavedrom==2.0.3.post3
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.3
